requirements.txt
certifi==2018.4.16
chardet==3.0.4
Django==2.0.6
django-cors-headers==2.1.0
django-filter==1.1.0
djangorestframework==3.7.7
idna==2.7
Markdown==2.6.11
mysqlclient==1.3.12
pytz==2018.3
requests==2.19.1
urllib3==1.23
